Send an extra float (h) when SPH_GRAV is on
Because only TBODYSZ bytes are communicated during tree traversal,
only the first 1+NDIM bytes of remote bodies were being
communicated... but gravity with SPH_GRAV on requires h in addition to
mass (1) and pos (NDIM).  Rearrange the body structure to put h right
after pos, and send the first 2+NDIM bytes if SPH_GRAV is on.