Force double alignment of SPHbody, add assertions to catch this
When some members of SPHbody are doubles and some are floats (for
example, when POS_IS_DOUBLE, which will almost always be true),
different platforms align structures differently, which can cause
structures to have different sizes due to padding.  On Intel Macs,
sizeof(SPHbody)/sizeof(float) was 57---i.e., it was not an even number
of doubles.  Since sizeof(double) couldn't divide evenly into
sizeof(SPHbody), SPHstride2 * sizeof(double) did not equal
sizeof(SPHbody).  SPHstride2 is used in the integration routines
(e.g., PUpdateXsd) to step through memory to the next particle, but
because the stride was wrong, the integrator went on a wild
data-corruption spree.

Added an explicit pad to SPHbody, to bring sizeof(SPHbody) to an even
number of doubles, and added assert statements near the top of main in
every application code, to hopefully catch this problem in the future
(when data members are added to, removed from, or rearranged within
SPHbody).

(Interestingly, without this explicit pad, sizeof(SPHbody) was already
an even number of doubles on my G5 mac and on x86_64 machines like
Coyote and Saguaro.  Adding the explicit pad to SPHbody shouldn't
change the size on those machines.)